refactor(app): the rest of models keeps shapes and hands values over
Four value maps and one map wedged into a form-union file move to
services/, where constant-maps puts a value — the codemirror one was
already importing the two services that build its entries, which is the
layering inversion that rule exists to stop. NamedItem is renamed for the
Item it exports, 3DResolver for the class every sibling names its file
after, and UnifiedColors reaches the vuetify config by alias.

The resolver and command class hierarchies stay models, like the Controls
classes. A type derived from a services map is a type-only import.

Closes the last three app/models units of the file-organization sweep.
