+++ packages/app/app/components/Message/Model/Status/PickerForm.vue
@@ -1,9 +1,9 @@
 <script setup lang="ts">
 import { SelectableStatusDefinitionList } from "@/models/message/user/status/SelectableStatusDefinitionList";
-import { StatusIconMap } from "@/models/message/user/status/StatusIconMap";
 import { MutationStatus } from "@/models/shared/MutationStatus";
 import { authClient } from "@/services/auth/authClient";
 import { StatusBadgePropsMap } from "@/services/message/StatusBadgePropsMap";
+import { StatusIconMap } from "@/services/message/user/status/StatusIconMap";
 import { useStatusStore } from "@/store/message/user/status";
 import { STATUS_MESSAGE_MAX_LENGTH } from "@esposter/db-schema";
 import { noop } from "@esposter/shared";
