+++ packages/app/app/components/Resource/Sheet/Column/EditDialog.vue
@@ -2,8 +2,9 @@
 import type { Column } from "#shared/models/resource/sheet/column/Column";
 import type { DataSource } from "#shared/models/resource/sheet/datasource/DataSource";
 
-import { columnFormSchema, ColumnTypeFormSchemaMap } from "@/models/resource/sheet/column/ColumnForm";
+import { columnFormSchema } from "@/models/resource/sheet/column/ColumnForm";
 import { zodToJsonSchema } from "@/services/jsonSchema/zodToJsonSchema";
+import { ColumnTypeFormSchemaMap } from "@/services/resource/sheet/column/ColumnTypeFormSchemaMap";
 import { getEditColumnDescription } from "@/services/resource/sheet/commands/getEditColumnDescription";
 import { extractSchemaFields } from "@/services/zod/extractSchemaFields";
 import { useColumnDialogStore } from "@/store/resource/sheet/columnDialog";
