+++ packages/app/app/models/colors/UnifiedColors.ts
@@ -1,5 +1,4 @@
+import type { BaseColors, getBaseColorsExtension } from "@@/vuetify.config";
 import type { Colors as BaseVuetifyColors } from "vuetify/lib/composables/theme.mjs";
-
-import type { BaseColors, getBaseColorsExtension } from "../../../vuetify.config";
 
 export type UnifiedColors = BaseColors & BaseVuetifyColors & ReturnType<typeof getBaseColorsExtension>;
