+++ packages/app/app/models/resource/ItemCategoryDefinition.ts
@@ -1,4 +1,4 @@
-import type { Item } from "@/models/entity/NamedItem";
+import type { Item } from "@/models/entity/Item";
 import type { ListItemCategoryDefinition } from "@/models/vuetify/ListItemCategoryDefinition";
 // Names the item field `value` is matched against, so a list of definitions can resolve the one
 // Describing any given item rather than only rendering a type in a vuetify select or dropdown
