+++ packages/app/app/models/resource/sheet/column/ColumnForm.ts
@@ -4,7 +4,6 @@ import type { DateColumnForm } from "@/models/resource/sheet/column/DateColumnFo
 import type { NumberColumnForm } from "@/models/resource/sheet/column/NumberColumnForm";
 import type { StringColumnForm } from "@/models/resource/sheet/column/StringColumnForm";
 
-import { ColumnType } from "#shared/models/resource/sheet/column/ColumnType";
 import { booleanColumnFormSchema } from "@/models/resource/sheet/column/BooleanColumnForm";
 import { computedColumnFormSchema } from "@/models/resource/sheet/column/ComputedColumnForm";
 import { dateColumnFormSchema } from "@/models/resource/sheet/column/DateColumnForm";
@@ -21,11 +20,3 @@ export const columnFormSchema = z.discriminatedUnion("type", [
   numberColumnFormSchema,
   stringColumnFormSchema,
 ]) satisfies z.ZodType<ColumnForm>;
-
-export const ColumnTypeFormSchemaMap = {
-  [ColumnType.Boolean]: booleanColumnFormSchema,
-  [ColumnType.Computed]: computedColumnFormSchema,
-  [ColumnType.Date]: dateColumnFormSchema,
-  [ColumnType.Number]: numberColumnFormSchema,
-  [ColumnType.String]: stringColumnFormSchema,
-} as const satisfies Record<ColumnType, z.ZodType<ColumnForm>>;
