+++ packages/app/app/services/codemirror/LanguageRegexSupportPatternMap.ts
