+++ packages/app/app/services/codemirror/getLanguage.ts
@@ -1,4 +1,4 @@
-import { LanguageRegexSupportPatternMap } from "@/models/codemirror/LanguageRegexSupportPatternMap";
+import { LanguageRegexSupportPatternMap } from "@/services/codemirror/LanguageRegexSupportPatternMap";
 
 export const getLanguage = (filename: string) => {
   for (const [language, supportPattern] of Object.entries(LanguageRegexSupportPatternMap))
