+++ packages/app/app/services/dashboard/chart/getActiveChartTypeResolvers.ts
@@ -2,8 +2,8 @@ import type { Chart } from "#shared/models/dashboard/data/chart/Chart";
 import type { ChartType } from "#shared/models/dashboard/data/chart/type/ChartType";
 import type { AChartTypeResolver } from "@/models/resolvers/dashboard/chart/AChartTypeResolver";
 
-import { ChartType3DResolver } from "@/models/resolvers/dashboard/chart/3DResolver";
 import { BasicResolver } from "@/models/resolvers/dashboard/chart/BasicResolver";
+import { ChartType3DResolver } from "@/models/resolvers/dashboard/chart/ChartType3DResolver";
 import { DistributedResolver } from "@/models/resolvers/dashboard/chart/DistributedResolver";
 import { DonutResolver } from "@/models/resolvers/dashboard/chart/DonutResolver";
 import { PyramidResolver } from "@/models/resolvers/dashboard/chart/PyramidResolver";
