+++ packages/app/app/services/message/FilterTypeHasIconMap.ts
